Сконфигурировать rate limiter в отдельном файле и импортировать в app.js.
